Added partial disabling of Teimo's store + Teimo's home bug fix + more items in minors object list
So first of all, I made a new Store class, that handles the Teimo's store. With it, the shop can be partially disabled, and then reneabled again, with the resuply, envelope, and Teimo's bike routine still working. So yeah, it can save couple of frames here and there. WIth that I added a new setting "Teimo's Shop".

I also fixed the issue with Teimo's house and chicken house (a big building on the road from Perjarvi to Ventti) form dissapearing when the Perajrvi is not loaded, by chanigng their parent from Perajarvi to null. I also added them to gameObjects list, so they're loaded while approached. It shouldn't cause any issues, but it needs testing.

I also added more objects to minor objects list (such as gas canister, rachet set, and few more).

And internally, I added new class called "Utilities" with "ContainsAny()" function. It allows to quickly check, if the string has any of the provided values in it.